Added RootBeer to android for better jailbreak detection
