+++ android/src/main/java/com/gantix/JailMonkey/ExternalStorage/ExternalStorageCheck.java
@@ -1,0 +1,42 @@
+package com.gantix.JailMonkey.ExternalStorage;
+
+import android.content.Context;
+import android.content.pm.ApplicationInfo;
+import android.content.pm.PackageInfo;
+import android.content.pm.PackageManager;
+import android.os.Build;
+
+public class ExternalStorageCheck {
+    /**
+     * Checks if the application is installed on the SD card.
+     *
+     * @return <code>true</code> if the application is installed on the sd card
+     */
+    public static boolean isOnExternalStorage(Context context) {
+        // check for API level 8 and higher
+        if (Build.VERSION.SDK_INT > android.os.Build.VERSION_CODES.ECLAIR_MR1) {
+            PackageManager pm = context.getPackageManager();
+            try {
+                PackageInfo pi = pm.getPackageInfo(context.getPackageName(), 0);
+                ApplicationInfo ai = pi.applicationInfo;
+                return (ai.flags & ApplicationInfo.FLAG_EXTERNAL_STORAGE) == ApplicationInfo.FLAG_EXTERNAL_STORAGE;
+            } catch (PackageManager.NameNotFoundException e) {
+                // ignore
+            }
+        }
+
+        // check for API level 7 - check files dir
+        try {
+            String filesDir = context.getFilesDir().getAbsolutePath();
+            if (filesDir.startsWith("/data/")) {
+                return false;
+            } else if (filesDir.contains("/mnt/") || filesDir.contains("/sdcard/")) {
+                return true;
+            }
+        } catch (Throwable e) {
+            // ignore
+        }
+
+        return false;
+    }
+}
