+++ android/src/main/java/com/gantix/JailMonkey/Rooted/CheckApiVersion.java
@@ -1,0 +1,6 @@
+package com.gantix.JailMonkey.Rooted;
+
+public interface CheckApiVersion {
+
+    boolean checkRooted();
+}
